add per-user settings; appropriate writeCharacteristic() for each user
Thanks to simon6162 on https://github.com/ec1oud/qt-ble-weightscale/issues/1
this became possible:

"This characteristic tells the scales how to process its readings and who
is selected in the app. So you need to set it before trying to read back
the data as you have now. This displays the p1, p2 etc and chooses the
display units. Also to calculate the body fat etc, it needs your age height
sex and activity level.

Assuming you are User 1 on your scales, Male, a low exerciser, 170 cm tall,
45 years old and report your weight in lbs then I believe this is the
definition of the characteristic.

TT = the header =0xfe
UI = user id (p1=0x01, p2=0x02 etc)
SS = Sex ( male = 0x01, female = 0x00)
EE = Exercise Level (0x00=ordinary, 0x01=amet, 0x02=professional)
HH = height in cm
AA = age in years (the app does this badly just on the year, form the look of it)
UU = Display units (04=st, lbs, 1=kg, 2=lbs)
CC = Checksum (UI^SS^EE^HH^AA^UU) - xor of the data

                    TT UI SS EE HH AA UU CC
1 Attribute Data: 0xfe 01 01 00 aa 2d 02 85
"

Now these settings are all stored via QSettings in separate sections.
The user name is still the primary key, thus the user ID can be changed.
Switching between users seems OK, like this: the user is not known until a
weight reading is received, but that can't happen until
writeCharacteristic(), which in turn depends on knowing which user it is.
So now I start by assuming the user is the same as last time; but if the
user is then found to be different, it will do the writeCharacteristic()
again, which hopefully will result in getting updated readings.
And it doesn't write to influxDB until it's sure it has readings for
the correct user.